Beginning
Remember to install the Win2d nuget package manager. Had some trouble when trying to use visuals with XAML. Right click on the solution explorer and click "restore nuGet packages" and I also right-clicked on the solution in the solution explorer and clicked configuration manager then used the x86 platform and it finally worked for me.